requirements.txt

Flask
azure-mgmt-resource
azure-mgmt-compute
azure-mgmt-network
azure-identity
azure-cli